+++ src/base/core/src/num/prob/rand/prng/lcg/mod.rs
@@ -1,0 +1,12 @@
+// devela_base_core::num::prob::rand::lcg
+//
+#![doc = concat![crate::_ABBR_LCG!(), "s."]]
+//
+
+mod u16;
+// mod u32;
+// mod u64;
+
+pub use u16::Lcg16;
+// pub use u32::Lcg32;
+// pub use u64::Lcg64;