+++ src/base/core/src/num/prob/rand/prng/mod.rs
@@ -9,7 +9,7 @@ mod shift; // XorShift[8|16|32|64|128], XorShift128p, xorshift_custom!
 #[cfg(feature = "rand")]
 crate::items! {
     #[cfg_attr(nightly_doc, doc(cfg(feature = "rand")))]
-    mod lgc; // Lgc16
+    mod lcg; // Lcg16
     #[cfg_attr(nightly_doc, doc(cfg(feature = "rand")))]
     mod xoroshiro; // Xoroshiro128pp
     #[cfg_attr(nightly_doc, doc(cfg(feature = "rand")))]
@@ -24,7 +24,7 @@ crate::structural_mods! { // _mods, _crate_internals
 
         #[cfg(feature = "rand")]
         pub use super::{
-            lgc::*,
+            lcg::*,
             xabc::*,
             xoroshiro::*,
             xyza8::*,