+++ docs/changelog.md
@@ -443,6 +443,7 @@ Many feature gates are removed in order to make most features make always availa
 
 #### rand
 - move `num::prob::rand` to [base].
+- rename `Lgc16` to `Lcg16`.
 - new traits: `Rand`, `RandAlloc`, `RandStd`.
 
 ### quant